summary_calculations now properly save their values in the database
Before, if you updated the value of a field which was being used
to calculated a calculated field with summary_calculation set to
true, the database wouldn't save the value if there was already
one written for this field at the 0th index.
record#update_calculated_fields had to be updated to destroy
the 0th indexed value first, so it could be saved in this case.