Added ability to specify that a condition should only apply to the zeroth index.
This is usefull for indexed presentations where you want to add constraints on to a non-indexed part of the form, i.e. create a non-indexed field that is dependent on some overall property of the the indexed fields that is determined by a condition.  In the simple_form spec see the has_mansion and is_mansion conditions and how they are used for an example.